+++ src/app/app.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import * as firebase from 'firebase/app';
+import { AngularFirestore } from 'angularfire2/firestore';
 
 @Component({
   selector: 'app-root',
@@ -7,15 +8,7 @@ import * as firebase from 'firebase/app';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-  constructor() {
-      const config = {
-      apiKey: "AIzaSyBMVhkrZj_9zrzpegDH5Dgy_G4CEvBJ1wY",
-      authDomain: "blogapp-5c679.firebaseapp.com",
-      databaseURL: "https://blogapp-5c679.firebaseio.com",
-      projectId: "blogapp-5c679",
-      storageBucket: "blogapp-5c679.appspot.com",
-      messagingSenderId: "724646466363"
-    };
-    firebase.initializeApp(config);
-  }
+
+  constructor(db: AngularFirestore) {}
+  
 }