+++ src/app/auth/login/login.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { ToasterService } from 'angular2-toaster';
 
 import { AuthService } from '../../services/auth.service';
 
@@ -17,7 +18,8 @@ export class LoginComponent implements OnInit {
   constructor(
     private formBuilder: FormBuilder,
     private router: Router,
-    private authService: AuthService
+    private authService: AuthService,
+    private toasterService: ToasterService
   ) { }
 
   ngOnInit(): void {
@@ -35,14 +37,57 @@ export class LoginComponent implements OnInit {
     const email = this.loginForm.get('email').value;
     const password = this.loginForm.get('password').value;
 
-    this.authService.authenticateUser(email, password).then(
+    this.authService.loginWithEmail(email, password).then(
       () => {
+        this.toasterService.pop('success', 'Success', 'You are now logged in');
         this.router.navigate(['/']);
       },
       (error) => {
         this.errorMessage = error;
       }
     );
+  }
+
+  loginWithFacebook(): void {
+    this.authService.loginWithFacebook().then(
+      () => {
+        this.toasterService.pop('success', 'Success', 'You are now logged in');
+        this.router.navigate(['/']);
+      },
+      (error) => {
+        this.errorMessage = error;
+      }
+    );
+  }
+
+  loginWithGoogle(): void {
+    this.authService.loginWithGoogle().then(
+      () => {
+        this.toasterService.pop('success', 'Success', 'You are now logged in');
+        this.router.navigate(['/']);
+      },
+      (error) => {
+        this.errorMessage = error;
+      }
+    );
+  }
+
+  onForgotPassword(): boolean {
+    const email = this.loginForm.get('email').value;
+
+    if(!email) {
+      this.errorMessage = 'Please enter your email address';
+    }
+    this.authService.sendPasswordResetEmail(email).then(
+      () => {
+        this.toasterService.pop('success', 'Email sent', 'Please check your mailbox');
+        this.router.navigate(['/']);
+      },
+      (error) => {
+        this.errorMessage = error;
+      }
+    );
+    return false;
   }
 
 }