+++ src/app/auth/register/register.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { ToasterService } from 'angular2-toaster';
 
 import { AuthService } from '../../services/auth.service';
 
@@ -17,7 +18,8 @@ export class RegisterComponent implements OnInit {
   constructor(
     private formBuilder: FormBuilder,
     private router: Router,
-    private authService: AuthService
+    private authService: AuthService,
+    private toasterService: ToasterService
   ) { }
 
   ngOnInit(): void {
@@ -26,17 +28,21 @@ export class RegisterComponent implements OnInit {
 
   initForm(): void {
     this.registerForm = this.formBuilder.group({
-      email: ['', [Validators.required, Validators.email]],
-      password: ['', [Validators.required, Validators.pattern(/[0-9a-zA-Z]{8,}/)]]
+      email: ['', [Validators.required, Validators.email]]
     });
   }
 
   onSubmit(): void {
     const email = this.registerForm.get('email').value;
-    const password = this.registerForm.get('password').value;
 
-    this.authService.registerUser(email, password).then(
+    var actionCodeSettings = {
+      url: window.location.href + '-confirm?email=' + email,
+      handleCodeInApp: true
+    };
+
+    this.authService.sendSignInLinkToEmail(email, actionCodeSettings).then(
       () => {
+        this.toasterService.pop('success', 'Email sent', 'Please check your mailbox');
         this.router.navigate(['/']);
       },
       (error) => {