+++ src/environments/environment.prod.ts
@@ -1,3 +1,11 @@
 export const environment = {
-  production: true
+  production: true,
+  firebase: {
+    apiKey: "AIzaSyBMVhkrZj_9zrzpegDH5Dgy_G4CEvBJ1wY",
+    authDomain: "blogapp-5c679.firebaseapp.com",
+    databaseURL: "https://blogapp-5c679.firebaseio.com",
+    projectId: "blogapp-5c679",
+    storageBucket: "blogapp-5c679.appspot.com",
+    messagingSenderId: "724646466363"
+  }
 };